✨ feat: base service

+++ src/controllers/tasks/tasks.controller.ts
@@ -16,28 +16,35 @@ export class TasksController {
     @GetUserId() userId: number,
     @Query() queryOptions?: { [key: string]: string },
   ) {
-    const tasks = this.tasksService.getAll(userId, queryOptions);
+    const tasks = this.tasksService.find({ owner: userId }, queryOptions);
     return tasks;
   }
 
   @Post('create')
   create(@GetUserId() userId: number, @Body() newTask: CreateTaskDto) {
-    const tasks = this.tasksService.create(userId, newTask);
+    const tasks = this.tasksService.create({
+      ...newTask,
+      completed: false,
+      owner: userId,
+    });
     return tasks;
   }
 
   @Delete()
   delete(@GetUserId() userId: number, @Body() { id }: DeleteTaskDto) {
-    const tasks = this.tasksService.delete(userId, id);
+    const tasks = this.tasksService.delete({ owner: userId, _id: id });
     return tasks;
   }
 
   @Post('edit')
   editTaskName(
     @GetUserId() userId: number,
-    @Body() taskInfos: EditTaskNameDto,
+    @Body() { id, newName }: EditTaskNameDto,
   ) {
-    const tasks = this.tasksService.editTaskName(userId, taskInfos);
+    const tasks = this.tasksService.updateOne(
+      { owner: userId, _id: id },
+      { name: newName },
+    );
     return tasks;
   }
 }
